+++ src/main/java/org/fogbowcloud/saps/engine/core/archiver/ArchiverHelper.java
@@ -13,7 +13,7 @@ import org.fogbowcloud.saps.engine.core.database.ImageDataStore;
 import org.fogbowcloud.saps.engine.core.model.ImageTask;
 import org.fogbowcloud.saps.engine.core.model.ImageTaskState;
 import org.fogbowcloud.saps.engine.core.util.CheckSumMD5ForFile;
-import org.fogbowcloud.saps.engine.scheduler.util.SapsPropertiesConstants;
+import org.fogbowcloud.saps.engine.util.SapsPropertiesConstants;
 import org.mapdb.DB;
 
 public class ArchiverHelper {
