+++ src/main/java/org/fogbowcloud/saps/engine/core/archiver/ftp/FTPIntegration.java
@@ -1,4 +1,4 @@
-package org.fogbowcloud.saps.engine.core.archiver;
+package org.fogbowcloud.saps.engine.core.archiver.ftp;
 
 import java.util.Properties;
