+++ src/main/java/org/fogbowcloud/saps/engine/core/scheduler/arrebol/exceptions/GetJobException.java
@@ -1,4 +1,4 @@
-package org.fogbowcloud.saps.engine.exceptions;
+package org.fogbowcloud.saps.engine.core.scheduler.arrebol.exceptions;
 
 public class GetJobException extends Throwable {
     public GetJobException(String s, Exception e) {
