+++ src/main/java/org/fogbowcloud/saps/engine/core/scheduler/arrebol/exceptions/SubmitJobException.java
@@ -1,4 +1,4 @@
-package org.fogbowcloud.saps.engine.exceptions;
+package org.fogbowcloud.saps.engine.core.scheduler.arrebol.exceptions;
 
 public class SubmitJobException extends Throwable {
     public SubmitJobException(String s, Exception e) {
