+++ src/main/java/org/fogbowcloud/saps/engine/restlet/resource/ProcessedImagesResource.java
@@ -1,8 +1,8 @@
-package org.fogbowcloud.saps.engine.scheduler.restlet.resource;
+package org.fogbowcloud.saps.engine.restlet.resource;
 
 import org.apache.commons.httpclient.HttpStatus;
 import org.apache.log4j.Logger;
-import org.fogbowcloud.saps.engine.scheduler.util.ProcessedImagesEmailBuilder;
+import org.fogbowcloud.saps.engine.util.ProcessedImagesEmailBuilder;
 import org.restlet.data.Form;
 import org.restlet.data.MediaType;
 import org.restlet.representation.Representation;
